// src/app/Models/course.ts
export interface Course {
  courseId: number;
  courseName: string;
  shortCode: string;
  totalHours: string;  
  courseFee: number;
  remarks: string;
  isActive: boolean;
  createdDate: Date;
}
//  courseId: number;
//  courseName: string;
//  shortCode: string;
//  durationMonths: string;
//  totalHours: string;
//  courseFee: number;
//  remarks: string;  
//  isActive: boolean;
//  createdDate: Date;

//}